+++ packages/cli/src/lib/constants.ts
@@ -2,3 +2,6 @@ export const LOGGER = 'logger';
 export const LOGGER_TRANSPORT = 'loggerTransport';
 export const COLLECTION = 'COLLECTION';
 export const META = 'META';
+
+export const DIRECTUS_COLLECTIONS_PREFIX = 'directus_';
+export const CUSTOM_COLLECTIONS_MAPPING_PREFIX = 'items:';
