--- packages/cli/src/lib/services/seed/constants.ts
@@ -1,2 +1,0 @@
-export const DIRECTUS_COLLECTIONS_PREFIX = 'directus_';
-export const CUSTOM_COLLECTIONS_PREFIX = 'items:';
